requirements.txt
tqdm
scipy
biopython
matplotlib
touch 
pandas 
numpy 
scikit-learn
tape_proteins